iommu/amd: Fix union of bitfields in intcapxt support
From 2fb6acf3edfeb904505f9ba3fd01166866062591 Mon Sep 17 00:00:00 2001
Git-commit: 2fb6acf3edfeb904505f9ba3fd01166866062591
Patch-mainline: v5.11-rc1
References: bsn#54

All the bitfields in here are overlaid on top of each other since
they're a union. Change the second u64 to be in a struct so it does
the intended thing.

Fixes: b5c3786ee370 ("iommu/amd: Use msi_msg shadow structs")
Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/r/20201111144322.1659970-2-dwmw2@infradead.org
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.11-rc1-iommu-amd-Fix-union-of-bitfields-in-intcapxt-support.patch