rcu: Add lockdep_assert_irqs_disabled() to
Git-commit: ea5e5bc881a4de9b5e09a16af950af2584dd67d2
 rcu_sched_clock_irq() and callees
References: bsn#19
Patch-mainline: v5.10.65

[ Upstream commit a649d25dcc671a33b9cc3176411920fdc5fbd98e ]

This commit adds a number of lockdep_assert_irqs_disabled() calls
to rcu_sched_clock_irq() and a number of the functions that it calls.
The point of this is to help track down a situation where lockdep appears
to be insisting that interrupts are enabled within these functions, which
should only ever be invoked from the scheduling-clock interrupt handler.

Link: https://lore.kernel.org/lkml/20201111133813.GA81547@elver.google.com/
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>
Signed-off-by: Sasha Levin <sashal@kernel.org>
Signed-off-by: Kai Liu <kai.liu@suse.com>

Patch-name: patches.stable/v5.10.65-0045-rcu-Add-lockdep_assert_irqs_disabled-to-rcu_sched_cl.patch