sched: Avoid unnecessary calculation of load imbalance at
From 5c339005f854fa75aa46078ad640919425658b3e Mon Sep 17 00:00:00 2001
 clone time
Git-commit: 5c339005f854fa75aa46078ad640919425658b3e
Patch-mainline: v5.11-rc1
References: bsn#12

In find_idlest_group(), the load imbalance is only relevant when the group
is either overloaded or fully busy but it is calculated unconditionally.
This patch moves the imbalance calculation to the context it is required.
Technically, it is a micro-optimisation but really the benefit is avoiding
confusing one type of imbalance with another depending on the group_type
in the next patch.

No functional change.

Signed-off-by: Mel Gorman <mgorman@techsingularity.net>
Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Reviewed-by: Vincent Guittot <vincent.guittot@linaro.org>
Link: https://lkml.kernel.org/r/20201120090630.3286-3-mgorman@techsingularity.net
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.11-rc1-sched-Avoid-unnecessary-calculation-of-load-imbalanc.patch