net: hns3: add support to utilize the firmware calculated
From e364ad303fe3e96ff30fb05c031774ecbbce4af1 Mon Sep 17 00:00:00 2001
 shaping parameters
Git-commit: e364ad303fe3e96ff30fb05c031774ecbbce4af1
Patch-mainline: v5.11-rc1
References: bsn#12

Since the calculation of the driver is fixed, if the number of
queue or clock changed, the calculated result may be inaccurate.

So for compatible and maintainable, add a new flag to tell the
firmware to calculate the shaping parameters with the specified
rate.

Signed-off-by: Yonglong Liu <liuyonglong@huawei.com>
Signed-off-by: Huazhong Tan <tanhuazhong@huawei.com>
Signed-off-by: Jakub Kicinski <kuba@kernel.org>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.11-rc1-net-hns3-add-support-to-utilize-the-firmware-calcula.patch