crypto: tcrypt - include 1420 byte blocks in aead and
From ad6d66bcac77e5145eb67449f8354ed0f936258c Mon Sep 17 00:00:00 2001
 skcipher benchmarks
Git-commit: ad6d66bcac77e5145eb67449f8354ed0f936258c
Patch-mainline: v5.11-rc1
References: resolve conflict when sync patch from OLK-5.10

WireGuard and IPsec both typically operate on input blocks that are
~1420 bytes in size, given the default Ethernet MTU of 1500 bytes and
the overhead of the VPN metadata.

Many aead and sckipher implementations are optimized for power-of-2
block sizes, and whether they perform well when operating on 1420
byte blocks cannot be easily extrapolated from the performance on
power-of-2 block size. So let's add 1420 bytes explicitly, and round
it up to the next blocksize multiple of the algo in question if it
does not support 1420 byte blocks.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.suse/v5.11-rc1-crypto-tcrypt-include-1420-byte-blocks-in-aead-and-s.patch