KVM: arm64: Refuse to run VCPU if PMU is not initialized
From 9bbfa4b565379eeb2fb8fdbcc9979549ae0e48d9 Mon Sep 17 00:00:00 2001
Git-commit: 9bbfa4b565379eeb2fb8fdbcc9979549ae0e48d9
Patch-mainline: v5.11-rc1
References: bsn#12

When enabling the PMU in kvm_arm_pmu_v3_enable(), KVM returns early if the
PMU flag created is false and skips any other checks. Because PMU emulation
is gated only on the VCPU feature being set, this makes it possible for
userspace to get away with setting the VCPU feature but not doing any
initialization for the PMU. Fix it by returning an error when trying to run
the VCPU if the PMU hasn't been initialized correctly.

The PMU is marked as created only if the interrupt ID has been set when
using an in-kernel irqchip. This means the same check in
kvm_arm_pmu_v3_enable() is redundant, remove it.

Signed-off-by: Alexandru Elisei <alexandru.elisei@arm.com>
Signed-off-by: Marc Zyngier <maz@kernel.org>
Link: https://lore.kernel.org/r/20201126144916.164075-1-alexandru.elisei@arm.com
Signed-off-by: Guoqing Jiang <guoqing.jiang@suse.com>

Patch-name: patches.mainline/v5.11-rc1-KVM-arm64-Refuse-to-run-VCPU-if-PMU-is-not-initializ.patch